Create com.apple.dock.json
Creation of com.apple.dock.json

json schema is unfinished/not fully tested. Need to add the NIST/CIS benchmark required keys to disable the ability to use hot-corners to disable screen saver activation or sleep. Also need to verify that the arrays for defining and configuring static items to be added to the dock are written/formatted properly an actually functional when deployed.